pyproject.toml
[tool.poetry]
name = "vistoria-back"
version = "0.1.0"
description = ""
authors = ["devclaudiocassimiro <claudio.silva@jala.university>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
reportlab = "^4.2.2"
flask-cors = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
